[REFACT] put model definition is separate py files
