+++ distribution_circuits_sale/models/__init__.py
@@ -2,3 +2,5 @@ from . import sale
 from . import time_frame
 from . import procurement
 from . import purchase
+from . import partner
+from . import product
