+++ distribution_circuits_sale/models/sale.py
@@ -67,69 +67,3 @@ class SaleOrderLine(models.Model):
         related='order_id.time_frame_id',
         store=True,
         string="Time Frame")
-
-
-class ProductTemplate(models.Model):
-    _inherit = "product.template"
-
-    @api.model
-    def _get_buy_route(self):
-        buy_route = self.env.ref('purchase.route_warehouse0_buy')
-        mto_route = self.env.ref('stock.route_warehouse0_mto')
-        routes_list = []
-
-        if buy_route:
-            routes_list.append(buy_route.ids[0])
-        if mto_route:
-            routes_list.append(mto_route.ids[0])
-        return routes_list
-
-    invoice_policy = fields.Selection(default='delivery')
-    type = fields.Selection(default='product')
-    route_ids = fields.Many2many(default=lambda self: self._get_buy_route())
-    uom_name = fields.Char(related="uom_id.name", string="UoM Name")
-    supplier_id = fields.Many2one(
-        compute="get_first_supplier",
-        comodel_name="res.partner",
-        string="Supplier")
-    supplier_name = fields.Char(related='supplier_id.display_name', string="Supplier Name")
-
-    @api.multi
-    def get_first_supplier(self):
-        for product in self:
-            for seller in product.sudo().seller_ids:
-                product.supplier_id = seller.name
-                break
-
-
-class Partner(models.Model):
-    _inherit = "res.partner"
-
-    amount_due = fields.Monetary(
-        string="Amount due for sale orders",
-        compute="_compute_amount_due")
-    customer_credit = fields.Monetary(
-        string="Customer credit",
-        compute="_compute_customer_credit")
-
-    @api.multi
-    def _compute_customer_credit(self):
-        for partner in self:
-            partner.customer_credit = -(partner.credit) - partner.amount_due
-
-    @api.multi
-    def _compute_amount_due(self):
-        order_obj = self.env['sale.order']
-
-        for partner in self:
-            orders = order_obj.search([
-                ('partner_id', '=', partner.id),
-                ('state', 'in', ['sent', 'sale']),
-                ('invoice_status', '!=', 'invoiced')
-                ])
-
-            amount_total = 0
-            for order in orders:
-                amount_total += order.amount_total
-
-            partner.amount_due = amount_total
